Initialize streams to be unbuffered
